Models for the PlayerElephant and a large lift
Currently the left leg bones are controlling all of the legs. May have to
separate the bones and re-add them; but the head already has some fixups
so that the face doesn't move much when the trunk moves, and these would
need to be redone.

The lift doors are a separate object from the lift, but both doors are
one object using a mirror modifier. Will see how it works after import.

SmoothShadedPlayerElephant.fbx is for finding out if the "smooth shading"
flag will automatically have an effect in Unity.

The PlayerElephant.blend file contains the source for the lift too,
I added it partly to be background for the scene.