Add floor sign and up/down arrows to the lift
Some objects and textures of interest:

There are now two doors, as separate objects.

There's a "Lift down sign" and "Lift up sign", with their own materials.

The floor sign has 3 materials on it. I'm expecting "Floor number foreground"
to be replaced with a transparent texture with the number written as text.